Split tests for executors into separate files.
