+++ tests/test_mp_graph_executor.py
@@ -43,15 +43,12 @@ from typing import TYPE_CHECKING, Any, Literal
 import networkx as nx
 import psutil
 
-from lsst.daf.butler.tests.utils import makeTestTempDir, removeTestTempDir
 from lsst.pipe.base import NodeId, QgraphSummary, QgraphTaskSummary
 from lsst.pipe.base.exec_fixup_data_id import ExecFixupDataId
 from lsst.pipe.base.graph import BuildId
 from lsst.pipe.base.mp_graph_executor import MPGraphExecutor, MPGraphExecutorError, MPTimeoutError
 from lsst.pipe.base.quantum_graph_executor import QuantumExecutor
 from lsst.pipe.base.quantum_reports import ExecutionStatus, QuantumReport, Report
-from lsst.pipe.base.single_quantum_executor import SingleQuantumExecutor
-from lsst.pipe.base.tests.simpleQGraph import AddTaskFactoryMock, makeSimpleQGraph
 
 if TYPE_CHECKING:
     import uuid
@@ -666,127 +663,6 @@ class MPGraphExecutorTestCase(unittest.TestCase):
         # Without DM-26728 fix the difference would be equal to number of
         # quanta (20).
         self.assertLess(num_fds_1 - num_fds_0, 5)
-
-
-class SingleQuantumExecutorTestCase(unittest.TestCase):
-    """Tests for SingleQuantumExecutor implementation."""
-
-    instrument = "lsst.pipe.base.tests.simpleQGraph.SimpleInstrument"
-
-    def setUp(self) -> None:
-        self.root = makeTestTempDir(TESTDIR)
-
-    def tearDown(self) -> None:
-        removeTestTempDir(self.root)
-
-    def test_simple_execute(self) -> None:
-        """Run execute() method in simplest setup."""
-        nQuanta = 1
-        butler, qgraph = makeSimpleQGraph(nQuanta, root=self.root, instrument=self.instrument)
-
-        nodes = list(qgraph)
-        self.assertEqual(len(nodes), nQuanta)
-        node = nodes[0]
-
-        task_factory = AddTaskFactoryMock()
-        executor = SingleQuantumExecutor(butler=butler, task_factory=task_factory)
-        executor.execute(node.task_node, node.quantum)
-        self.assertEqual(task_factory.countExec, 1)
-
-        # There must be one dataset of task's output connection
-        refs = list(butler.registry.queryDatasets("add_dataset1", collections=butler.run))
-        self.assertEqual(len(refs), 1)
-
-    def test_skip_existing_execute(self) -> None:
-        """Run execute() method twice, with skip_existing_in."""
-        nQuanta = 1
-        butler, qgraph = makeSimpleQGraph(nQuanta, root=self.root, instrument=self.instrument)
-
-        nodes = list(qgraph)
-        self.assertEqual(len(nodes), nQuanta)
-        node = nodes[0]
-
-        task_factory = AddTaskFactoryMock()
-        executor = SingleQuantumExecutor(butler=butler, task_factory=task_factory)
-        executor.execute(node.task_node, node.quantum)
-        self.assertEqual(task_factory.countExec, 1)
-
-        refs = list(butler.registry.queryDatasets("add_dataset1", collections=butler.run))
-        self.assertEqual(len(refs), 1)
-        dataset_id_1 = refs[0].id
-
-        # Re-run it with skip_existing_in, it should not run.
-        assert butler.run is not None
-        executor = SingleQuantumExecutor(
-            butler=butler, task_factory=task_factory, skip_existing_in=[butler.run]
-        )
-        executor.execute(node.task_node, node.quantum)
-        self.assertEqual(task_factory.countExec, 1)
-
-        refs = list(butler.registry.queryDatasets("add_dataset1", collections=butler.run))
-        self.assertEqual(len(refs), 1)
-        dataset_id_2 = refs[0].id
-        self.assertEqual(dataset_id_1, dataset_id_2)
-
-    def test_clobber_outputs_execute(self) -> None:
-        """Run execute() method twice, with clobber_outputs."""
-        nQuanta = 1
-        butler, qgraph = makeSimpleQGraph(nQuanta, root=self.root, instrument=self.instrument)
-
-        nodes = list(qgraph)
-        self.assertEqual(len(nodes), nQuanta)
-        node = nodes[0]
-
-        task_factory = AddTaskFactoryMock()
-        executor = SingleQuantumExecutor(butler=butler, task_factory=task_factory)
-        executor.execute(node.task_node, node.quantum)
-        self.assertEqual(task_factory.countExec, 1)
-
-        refs = list(butler.registry.queryDatasets("add_dataset1", collections=butler.run))
-        self.assertEqual(len(refs), 1)
-        dataset_id_1 = refs[0].id
-
-        original_dataset = butler.get(refs[0])
-
-        # Remove the dataset ourself, and replace it with something
-        # different so we can check later whether it got replaced.
-        butler.pruneDatasets([refs[0]], disassociate=False, unstore=True, purge=False)
-        replacement = original_dataset + 10
-        butler.put(replacement, refs[0])
-
-        # Re-run it with clobber_outputs and skip_existing_in, it should not
-        # clobber but should skip instead.
-        assert butler.run is not None
-        executor = SingleQuantumExecutor(
-            butler=butler, task_factory=task_factory, skip_existing_in=[butler.run], clobber_outputs=True
-        )
-        executor.execute(node.task_node, node.quantum)
-        self.assertEqual(task_factory.countExec, 1)
-
-        refs = list(butler.registry.queryDatasets("add_dataset1", collections=butler.run))
-        self.assertEqual(len(refs), 1)
-        dataset_id_2 = refs[0].id
-        self.assertEqual(dataset_id_1, dataset_id_2)
-
-        second_dataset = butler.get(refs[0])
-        self.assertEqual(list(second_dataset), list(replacement))
-
-        # Re-run it with clobber_outputs but without skip_existing_in, it
-        # should clobber.
-        assert butler.run is not None
-        executor = SingleQuantumExecutor(butler=butler, task_factory=task_factory, clobber_outputs=True)
-        executor.execute(node.task_node, node.quantum)
-        self.assertEqual(task_factory.countExec, 2)
-
-        refs = list(butler.registry.queryDatasets("add_dataset1", collections=butler.run))
-        self.assertEqual(len(refs), 1)
-        dataset_id_3 = refs[0].id
-
-        third_dataset = butler.get(refs[0])
-        self.assertEqual(list(third_dataset), list(original_dataset))
-
-        # No change in UUID even after replacement
-        self.assertEqual(dataset_id_1, dataset_id_3)
 
 
 def setup_module(module):
